chore: move code into a modules folder
