+++ src/modules/exceptions/uncaught.exception.ts
