+++ src/modules/exceptions/unhandled-promise.exception.ts
